+++ src/main/kotlin/r2dbcfun/R2dbcRepo.kt
@@ -1,24 +1,14 @@
 package r2dbcfun
 
-import io.r2dbc.spi.Clob
 import io.r2dbc.spi.Connection
-import io.r2dbc.spi.Statement
 import kotlinx.coroutines.flow.Flow
-import kotlinx.coroutines.flow.collect
-import kotlinx.coroutines.flow.map
 import kotlinx.coroutines.flow.single
-import kotlinx.coroutines.reactive.asFlow
-import kotlinx.coroutines.reactive.awaitSingle
 import r2dbcfun.internal.IDHandler
-import java.lang.Enum.valueOf
 import kotlin.reflect.KClass
 import kotlin.reflect.KFunction
-import kotlin.reflect.KParameter
 import kotlin.reflect.KProperty1
 import kotlin.reflect.full.declaredMemberProperties
-import kotlin.reflect.full.isSubclassOf
 import kotlin.reflect.full.primaryConstructor
-import kotlin.reflect.jvm.javaType
 
 public interface PK {
     public val id: Long
@@ -51,150 +41,9 @@ public class R2dbcRepo<T : Any, PKClass : PK>(
     @Suppress("UNCHECKED_CAST")
     private val idProperty = properties["id"] as KProperty1<T, Any>
 
-    private class Inserter<T : Any, PKClass : PK>(
-        table: String,
-        val connection: Connection,
-        val insertProperties: ArrayList<KProperty1<T, *>>,
-        val idHandler: IDHandler<T, PKClass>
-    ) {
-        private val insertStatementString = run {
-            val fieldNames = insertProperties.joinToString { it.name.toSnakeCase() }
-            val fieldPlaceHolders = (1..insertProperties.size).joinToString { idx -> "$$idx" }
-            "INSERT INTO $table($fieldNames) values ($fieldPlaceHolders)"
-        }
-
-        suspend fun create(instance: T): T {
-            val statement = insertProperties.foldIndexed(
-                connection.createStatement(insertStatementString)
-            )
-            { idx, statement, property ->
-                bindValueOrNull(
-                    statement,
-                    idx,
-                    property.call(instance),
-                    property.returnType.classifier as KClass<*>,
-                    property.name
-                )
-            }
-            val id = try {
-                statement.executeInsert()
-            } catch (e: Exception) {
-                throw R2dbcRepoException("error executing insert: $insertStatementString", e)
-            }
-
-            return idHandler.assignId(instance, id)
-        }
-    }
-
     private val inserter = Inserter(tableName, connection, propertiesExceptId, idAssigner)
 
-    private class Updater<T : Any, PKClass : PK>(
-        table: String,
-        val connection: Connection,
-        val updateProperties: ArrayList<KProperty1<T, *>>,
-        val idHandler: IDHandler<T, PKClass>,
-        val id: KProperty1<T, Any>
-    ) {
-        private val updateStatementString = run {
-            val propertiesString = updateProperties.withIndex()
-                .joinToString { indexedProperty -> "${indexedProperty.value.name.toSnakeCase()}=$${indexedProperty.index + 2}" }
-
-            @Suppress("SqlResolve")
-            "UPDATE $table set $propertiesString where id=$1"
-        }
-
-        suspend fun update(instance: T) {
-            val statement = updateProperties.foldIndexed(
-                connection.createStatement(updateStatementString)
-                    .bind(0, idHandler.getId(id.call(instance)))
-            ) { idx, statement, entry ->
-                bindValueOrNull(
-                    statement,
-                    idx + 1,
-                    entry.call(instance),
-                    entry.returnType.classifier as KClass<*>,
-                    entry.name
-                )
-            }
-            val rowsUpdated = statement.execute().awaitSingle().rowsUpdated.awaitSingle()
-            if (rowsUpdated != 1)
-                throw R2dbcRepoException("rowsUpdated was $rowsUpdated instead of 1")
-
-        }
-
-    }
-
     private val updater = Updater(tableName, connection, propertiesExceptId, idAssigner, idProperty)
-
-    private class Finder<T : Any, PKClass : PK>(
-        val table: String,
-        val connection: Connection,
-        val idHandler: IDHandler<T, PKClass>,
-        val constructor: KFunction<T>,
-        kClass: KClass<T>
-    ) {
-        @Suppress("SqlResolve")
-        private val selectString =
-            "select ${constructor.parameters.joinToString { it.name!!.toSnakeCase() }} from $table where "
-        private val snakeCaseStringForConstructorParameter =
-            constructor.parameters.associateBy({ it }, { it.name!!.toSnakeCase() })
-        private val snakeCaseForProperty =
-            kClass.declaredMemberProperties.associateBy({ it }, { it.name.toSnakeCase() })
-
-
-        suspend fun <V> findBy(property: KProperty1<T, V>, propertyValue: V): Flow<T> {
-            val query = selectString + snakeCaseForProperty[property] + "=$1"
-            val queryResult = try {
-                connection.createStatement(query).bind("$1", propertyValue).execute()
-                    .awaitSingle()
-            } catch (e: Exception) {
-                throw R2dbcRepoException("error executing select: $query", e)
-            }
-            val parameters = queryResult.map { row, _ ->
-                snakeCaseStringForConstructorParameter.mapValues { entry ->
-                    row.get(entry.value)
-                }
-            }.asFlow()
-            return parameters.map {
-                val resolvedParameters: Map<KParameter, Any> = it.mapValues { (parameter, value) ->
-                    val resolvedValue = when (value) {
-                        is Clob -> {
-                            val sb = StringBuilder()
-                            value.stream().asFlow().collect { chunk ->
-                                @Suppress("BlockingMethodInNonBlockingContext")
-                                sb.append(chunk)
-                            }
-                            value.discard()
-                            sb.toString()
-                        }
-                        else -> value
-                    }
-                    if (parameter.name == "id")
-                        idHandler.createId(resolvedValue as Long)
-                    else {
-
-                        val clazz = parameter.type.javaType as Class<*>
-                        if (resolvedValue != null && clazz.isEnum) {
-                            createEnumValue(clazz, resolvedValue)
-                        } else
-                            resolvedValue
-                    }
-                }
-                try {
-                    constructor.callBy(resolvedParameters)
-                } catch (e: IllegalArgumentException) {
-                    throw R2dbcRepoException(
-                        "error invoking constructor for $table. parameters:$resolvedParameters",
-                        e
-                    )
-                }
-            }
-        }
-
-        private fun createEnumValue(clazz: Class<*>, resolvedValue: Any?) =
-            (@Suppress("UPPER_BOUND_VIOLATED", "UNCHECKED_CAST")
-            valueOf<Any>(clazz as Class<Any>, resolvedValue as String))
-    }
 
     private val finder = Finder(tableName, connection, idAssigner, constr, kClass)
 
@@ -234,25 +83,3 @@ public class R2dbcRepo<T : Any, PKClass : PK>(
 
 }
 
-private fun bindValueOrNull(
-    statement: Statement,
-    index: Int,
-    value: Any?,
-    kClass: KClass<*>,
-    fieldName: String
-): Statement {
-    return try {
-        if (value == null) {
-            val clazz = if (kClass.isSubclassOf(Enum::class)) String::class.java else kClass.java
-            statement.bindNull(index, clazz)
-        } else {
-
-            statement.bind(index, if (value::class.isSubclassOf(Enum::class)) value.toString() else value)
-        }
-    } catch (e: java.lang.IllegalArgumentException) {
-        throw R2dbcRepoException(
-            "error binding value $value to field $fieldName with index $index",
-            e
-        )
-    }
-}
