+++ src/main/kotlin/r2dbcfun/Updater.kt
@@ -1,0 +1,43 @@
+package r2dbcfun
+
+import io.r2dbc.spi.Connection
+import kotlinx.coroutines.reactive.awaitSingle
+import r2dbcfun.internal.IDHandler
+import kotlin.reflect.KClass
+import kotlin.reflect.KProperty1
+
+internal class Updater<T : Any, PKClass : PK>(
+    table: String,
+    val connection: Connection,
+    val updateProperties: ArrayList<KProperty1<T, *>>,
+    val idHandler: IDHandler<T, PKClass>,
+    val id: KProperty1<T, Any>
+) {
+    private val updateStatementString = run {
+        val propertiesString = updateProperties.withIndex()
+            .joinToString { indexedProperty -> "${indexedProperty.value.name.toSnakeCase()}=$${indexedProperty.index + 2}" }
+
+        @Suppress("SqlResolve")
+        "UPDATE $table set $propertiesString where id=$1"
+    }
+
+    suspend fun update(instance: T) {
+        val statement = updateProperties.foldIndexed(
+            connection.createStatement(updateStatementString)
+                .bind(0, idHandler.getId(id.call(instance)))
+        ) { idx, statement, entry ->
+            bindValueOrNull(
+                statement,
+                idx + 1,
+                entry.call(instance),
+                entry.returnType.classifier as KClass<*>,
+                entry.name
+            )
+        }
+        val rowsUpdated = statement.execute().awaitSingle().rowsUpdated.awaitSingle()
+        if (rowsUpdated != 1)
+            throw R2dbcRepoException("rowsUpdated was $rowsUpdated instead of 1")
+
+    }
+
+}
